user misconfigurations in soda checks are masked
User misconfigurations for scans are not reported to the user and results are published regardless.
With the latest option to send slack notifications on passed soda checks we will mask
a potential configuration error, e.g. a check referrering to a
nonexisting column in a table.

Proposing with this PR to log error and not publish the scan results if
the user has misconfigurations leading to query execution errors.